+++ packages/core/src/components.d.ts
@@ -26,7 +26,7 @@ import { FocusTargetValueOrFalse } from "focus-trap";
 import { OzonContentAnchorClick, OzonContentClick } from "./components/ozon-content/ozon-content.interfaces";
 import { PaginationSelectPageEvent } from "./components/pagination/pagination.interfaces";
 import { ResponsiveElementSize } from "./components/responsive-element/responsive-element.interfaces";
-import { DsoScrollContainerEvent } from "./components/scroll-container/scroll-container.interfaces";
+import { DsoScrollableEvent } from "./components/scrollable/scrollable.interfaces";
 import { SelectableChangeEvent } from "./components/selectable/selectable.interfaces";
 import { SlideToggleActiveEvent } from "./components/slide-toggle/slide-toggle.interfaces";
 import { TreeViewItem, TreeViewPointerEvent } from "./components/tree-view/tree-view.interfaces";
@@ -414,7 +414,7 @@ export namespace Components {
     interface DsoResponsiveElement {
         "getSize": () => Promise<ResponsiveElementSize>;
     }
-    interface DsoScrollContainer {
+    interface DsoScrollable {
         /**
           * Internal method. Do not use.
          */
@@ -602,9 +602,9 @@ export interface DsoResponsiveElementCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLDsoResponsiveElementElement;
 }
-export interface DsoScrollContainerCustomEvent<T> extends CustomEvent<T> {
+export interface DsoScrollableCustomEvent<T> extends CustomEvent<T> {
     detail: T;
-    target: HTMLDsoScrollContainerElement;
+    target: HTMLDsoScrollableElement;
 }
 export interface DsoSelectableCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -833,11 +833,11 @@ declare global {
         prototype: HTMLDsoResponsiveElementElement;
         new (): HTMLDsoResponsiveElementElement;
     };
-    interface HTMLDsoScrollContainerElement extends Components.DsoScrollContainer, HTMLStencilElement {
+    interface HTMLDsoScrollableElement extends Components.DsoScrollable, HTMLStencilElement {
     }
-    var HTMLDsoScrollContainerElement: {
-        prototype: HTMLDsoScrollContainerElement;
-        new (): HTMLDsoScrollContainerElement;
+    var HTMLDsoScrollableElement: {
+        prototype: HTMLDsoScrollableElement;
+        new (): HTMLDsoScrollableElement;
     };
     interface HTMLDsoSelectableElement extends Components.DsoSelectable, HTMLStencilElement {
     }
@@ -917,7 +917,7 @@ declare global {
         "dso-progress-bar": HTMLDsoProgressBarElement;
         "dso-progress-indicator": HTMLDsoProgressIndicatorElement;
         "dso-responsive-element": HTMLDsoResponsiveElementElement;
-        "dso-scroll-container": HTMLDsoScrollContainerElement;
+        "dso-scrollable": HTMLDsoScrollableElement;
         "dso-selectable": HTMLDsoSelectableElement;
         "dso-slide-toggle": HTMLDsoSlideToggleElement;
         "dso-table": HTMLDsoTableElement;
@@ -1341,11 +1341,11 @@ declare namespace LocalJSX {
     interface DsoResponsiveElement {
         "onDsoSizeChange"?: (event: DsoResponsiveElementCustomEvent<ResponsiveElementSize>) => void;
     }
-    interface DsoScrollContainer {
+    interface DsoScrollable {
         /**
           * Event emitted when the scrollbar has reached top or bottom.
          */
-        "onDsoScrollContainerEvent"?: (event: DsoScrollContainerCustomEvent<DsoScrollContainerEvent>) => void;
+        "onDsoScrollableEvent"?: (event: DsoScrollableCustomEvent<DsoScrollableEvent>) => void;
     }
     interface DsoSelectable {
         "checked"?: boolean;
@@ -1491,7 +1491,7 @@ declare namespace LocalJSX {
         "dso-progress-bar": DsoProgressBar;
         "dso-progress-indicator": DsoProgressIndicator;
         "dso-responsive-element": DsoResponsiveElement;
-        "dso-scroll-container": DsoScrollContainer;
+        "dso-scrollable": DsoScrollable;
         "dso-selectable": DsoSelectable;
         "dso-slide-toggle": DsoSlideToggle;
         "dso-table": DsoTable;
@@ -1540,7 +1540,7 @@ declare module "@stencil/core" {
             "dso-progress-bar": LocalJSX.DsoProgressBar & JSXBase.HTMLAttributes<HTMLDsoProgressBarElement>;
             "dso-progress-indicator": LocalJSX.DsoProgressIndicator & JSXBase.HTMLAttributes<HTMLDsoProgressIndicatorElement>;
             "dso-responsive-element": LocalJSX.DsoResponsiveElement & JSXBase.HTMLAttributes<HTMLDsoResponsiveElementElement>;
-            "dso-scroll-container": LocalJSX.DsoScrollContainer & JSXBase.HTMLAttributes<HTMLDsoScrollContainerElement>;
+            "dso-scrollable": LocalJSX.DsoScrollable & JSXBase.HTMLAttributes<HTMLDsoScrollableElement>;
             "dso-selectable": LocalJSX.DsoSelectable & JSXBase.HTMLAttributes<HTMLDsoSelectableElement>;
             "dso-slide-toggle": LocalJSX.DsoSlideToggle & JSXBase.HTMLAttributes<HTMLDsoSlideToggleElement>;
             "dso-table": LocalJSX.DsoTable & JSXBase.HTMLAttributes<HTMLDsoTableElement>;