+++ packages/core/src/components/scroll-container/scroll-container.interfaces.ts
@@ -1,3 +1,0 @@
-export type DsoScrollContainerEvent = {
-  scrollEnd: "top" | "bottom";
-};