+++ packages/core/src/components/scrollable/scrollable.interfaces.ts
@@ -1,0 +1,3 @@
+export type DsoScrollableEvent = {
+  scrollEnd: "top" | "bottom";
+};