+++ packages/dso-toolkit/src/components/scroll-container/scroll-container.models.ts
@@ -1,8 +1,0 @@
-export interface ScrollContainer<TemplateFnReturnType> {
-  dsoScrollContainerEvent?: (e: CustomEvent<DsoScrollContainerEvent>) => void;
-  children: TemplateFnReturnType;
-}
-
-export type DsoScrollContainerEvent = {
-  scrollEnd: "top" | "bottom" | false;
-};