+++ packages/dso-toolkit/src/components/scrollable/scrollable.models.ts
@@ -1,0 +1,8 @@
+export interface Scrollable<TemplateFnReturnType> {
+  dsoScrollableEvent?: (e: CustomEvent<DsoScrollableEvent>) => void;
+  children: TemplateFnReturnType;
+}
+
+export type DsoScrollableEvent = {
+  scrollEnd: "top" | "bottom" | false;
+};