+++ packages/react/src/components/scrollable/scrollable.decorator.tsx
@@ -2,7 +2,7 @@ import { DecoratorFunction } from "@storybook/addons";
 import React from "react";
 
 export const decorator: DecoratorFunction<JSX.Element> = (story) => (
-  <div id="scroll-container-mock" style={{ backgroundColor: "#efefef", height: "750px", width: "500px" }}>
+  <div id="scrollable-mock" style={{ backgroundColor: "#efefef", height: "750px", width: "500px" }}>
     {story()}
   </div>
 );