+++ packages/react/src/templates.ts
@@ -32,7 +32,7 @@ import {
   ProgressBar,
   ProgressIndicator,
   ResponsiveElement,
-  ScrollContainer,
+  Scrollable,
   Selectable,
   SlideToggle,
   Table,
@@ -74,7 +74,7 @@ import { reactPagination } from "./components/pagination/pagination.react-templa
 import { reactProgressBar } from "./components/progress-bar/progress-bar.react-template";
 import { reactProgressIndicator } from "./components/progress-indicator/progress-indicator.react-template";
 import { reactResponsiveElement } from "./components/responsive-element/responsive-element.react-template";
-import { reactScrollContainer } from "./components/scroll-container/scroll-container.react-template";
+import { reactScrollable } from "./components/scrollable/scrollable.react-template";
 import { reactSelectable } from "./components/selectable/selectable.react-template";
 import { reactSlideToggle } from "./components/slide-toggle/slide-toggle.react-template";
 import { reactTable } from "./components/table/table.react-template";
@@ -115,7 +115,7 @@ export interface Components {
   progressBar: ProgressBar;
   progressIndicator: ProgressIndicator;
   responsiveElement: ResponsiveElement<JSX.Element>;
-  scrollContainer: ScrollContainer<JSX.Element>;
+  scrollable: Scrollable<JSX.Element>;
   selectable: Selectable<JSX.Element>;
   slideToggle: SlideToggle;
   table: Table<JSX.Element>;
@@ -162,7 +162,7 @@ templateContainer.add(reactPagination);
 templateContainer.add(reactProgressBar);
 templateContainer.add(reactProgressIndicator);
 templateContainer.add(reactResponsiveElement);
-templateContainer.add(reactScrollContainer);
+templateContainer.add(reactScrollable);
 templateContainer.add(reactSelectable);
 templateContainer.add(reactSlideToggle);
 templateContainer.add(reactTable);